kfs: task: made pid u16 and created pid_bitset
Task are using pid_bitset to alloc new pids instead of single integer.
Pid is u16 now instead of u32.
Fixed bug in setPGID() syscall

+++ src/debug/mm.zig
@@ -245,7 +245,7 @@ pub fn printMapped() void {
 }
 
 
-pub fn printTaskVMAs(pid: u32) void {
+pub fn printTaskVMAs(pid: u16) void {
     if (krn.task.initial_task.findByPid(pid)) |tsk| {
         if (tsk.mm.?.vmas) |vma| {
             var it = vma.list.iterator();
